@@ -56,7 +56,7 @@ class ConnectionID extends QuicID {
   }
 
   static fromString (str) {
-    return new ConnectionID(new Buffer(str, 'hex'))
+    return new ConnectionID(Buffer.from(str, 'hex'))
   }
 
   static random () {
