@@ -412,7 +412,7 @@ class AckFrame extends QuicFrame {
     if (buf.length < v.end) throw new QuicError('INVALID_FRAME_DATA')
     let ackBlockLength = buf.readUIntLE(v.start, missingSequenceNumberDeltaLen, true)
     let largestAcked = frame.largestAcked.value
-    if (largestAcked > 0 && ackBlockLength < 1 || (ackBlockLength > largestAcked)) {
+    if (largestAcked > 0 && (ackBlockLength < 1 || (ackBlockLength > largestAcked))) {
       throw new QuicError('INVALID_FRAME_DATA')
     }
 
